[MISC] SAM/BAM restructering. Puts shared functionality of SAM/BAM into a format_base class, define large functions outside of the class bodies.
